+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Features
+- Added indicator of current ship in 'armada info'.
+
+### Bug fixes
+- Fix log file permission error. 
+- Validate response from armada_api.get('version').
+
+
 ## 0.20.2 (2016-07-04)
 
 We do best effort to support docker versions 1.6.0 - 1.10.3 with this release.